BREAKING CHANGE: rename user_context tool to shortcuts_usage
- Renamed tool from user_context to shortcuts_usage for clarity
- Renamed source files from user-context.* to shortcuts-usage.*
- Updated all imports and references throughout codebase
- Added note in tool description about previous name
- Updated README and manifest.json

Users must update their Claude Desktop configuration to use the new tool name.
